# docs/API.yaml
openapi: 3.0.3
info:
  title: Tasky API
  description: |
    REST API for Tasky — Mongolia's most trusted domestic service marketplace.

    ## Authentication
    All endpoints (unless noted) require a Bearer JWT in the `Authorization` header.
    Obtain tokens via the OTP login flow (`/auth/otp/*`).

    ## Pagination
    All list endpoints use **cursor-based pagination**.
    Pass `cursor` (opaque string) and `limit` (default 20, max 100) as query parameters.
    Responses wrap results in `{ data: [], cursor: { next, has_more } }`.

    ## Errors
    All errors return a standard JSON body: `{ code, message, trace_id }`.

    ## Currency
    All monetary amounts are integers in MNT (Mongolian Tugrik).

    ## Idempotency
    Critical state-changing endpoints require an `Idempotency-Key` header.
    Payment and wallet endpoints are documented as post-MVP deferred surfaces.
  version: 1.1.0
  contact:
    name: Tasky Engineering
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CursorParam:
      name: cursor
      in: query
      description: Opaque cursor for pagination. Pass the `next` value from a previous response.
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      description: Maximum number of items to return (default 20, max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Unique key to ensure idempotent handling of critical state-changing requests.
  schemas:
...
    Error:
      type: object
      required: [code, message, trace_id]
      properties:
        code:
          type: string
          example: TASK_ALREADY_BOOKED
        message:
          type: string
          example: This task has already been assigned to another tasker.
        trace_id:
          type: string
          example: abc-123-def
    CursorPagination:
      type: object
      required: [next, has_more]
      properties:
        next:
          type: string
          nullable: true
          description: Opaque cursor to fetch the next page. Null if no more results.
          example: eyJpZCI6MTAwfQ==
        has_more:
          type: boolean
          example: true
...
    WalletBalance:
      type: object
      required: [balance, pending_payout, currency]
      properties:
        balance:
          type: integer
          description: Available balance in MNT.
          example: 250000
        pending_payout:
          type: integer
          description: Amount currently requested for payout but not yet processed.
          example: 100000
        currency:
          type: string
          example: MNT
    LedgerEntry:
      type: object
      required: [id, amount, type, description, created_at]
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: integer
          description: Signed amount in MNT. Positive = credit, negative = debit.
        type:
          type: string
          enum: [DEPOSIT, FEE, PAYOUT, REFUND]
        reference_id:
          type: string
          format: uuid
          nullable: true
          description: Related booking or payout ID.
        description:
          type: string
          example: "Цэвэрлэгээ ажил #1234 — орлого"
        created_at:
          type: string
          format: date-time
    PayoutRequest:
      type: object
      required: [id, amount, bank_name, bank_account, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        amount:
          type: integer
          description: Payout amount in MNT.
        bank_name:
          type: string
          example: Хаан банк
        bank_account:
          type: string
          example: "5012345678"
        status:
          type: string
          enum: [PENDING, PROCESSED, REJECTED]
        created_at:
          type: string
          format: date-time
        processed_at:
          type: string
          format: date-time
          nullable: true

    # ─── Reviews ──────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: Invalid request parameters or body.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid authentication token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /wallet:
    get:
      tags: [Wallet]
      summary: Get wallet balance
      deprecated: true
      description: Post-MVP deferred endpoint. Returns wallet balance and pending payout amount.
      operationId: getWalletBalance
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Wallet balance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletBalance'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: Monetization is deferred for the current MVP phase.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /wallet/transactions:
    get:
      tags: [Wallet]
      summary: List wallet transactions
      deprecated: true
      description: Post-MVP deferred endpoint. Returns paginated transaction history.
      operationId: listWalletTransactions
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated transaction history.
          content:
            application/json:
              schema:
                type: object
                required: [data, cursor]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LedgerEntry'
                  cursor:
                    $ref: '#/components/schemas/CursorPagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: Monetization is deferred for the current MVP phase.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /wallet/payouts:
    post:
      tags: [Wallet]
      summary: Request payout
      deprecated: true
      description: |
        Post-MVP deferred endpoint.
        Tasker requests a payout from their wallet balance to a bank account.
        Payouts are processed by admin on Tuesdays and Fridays.
        Requires Idempotency-Key header.
      operationId: requestPayout
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, bank_name, bank_account]
              properties:
                amount:
                  type: integer
                  minimum: 10000
                  description: Payout amount in MNT. Minimum 10,000 MNT.
                bank_name:
                  type: string
                  example: Хаан банк
                bank_account:
                  type: string
                  example: "5012345678"
      responses:
        '201':
          description: Payout request created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayoutRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Insufficient balance or existing pending payout.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Monetization is deferred for the current MVP phase.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ─── Reviews ────────────────────────────────────────────────
...